Optimize Camera worker code
Move db queries out of the loop as much as possible to avoid hitting the database when the camera workers is polling for every second.

Renamed all instances of camera_id where it actually meant camera_exid

Disabled checking S3 for the presence of file before updating the database with thumbnail url. This reduces 1-2 http calls for every single camera upload.